FuelLevel(Draft)
Included Fuel Guage files and damping filter files draft version

+++ STM32CubeIDE/Application/User/Core/src/Service/System_Service/IgnitionManager/IGN_SmHandler.c
@@ -230,7 +230,7 @@ void vInIgnitionOffStateActionFunc(void)
  */
 uint16_t usIgnitionGetCurrentState(void)
 {
-    return(SM_get_current_state(IGN_SM_INDEX));
+    return((uint16_t)SM_get_current_state(IGN_SM_INDEX));
 
 }
